Scoring added + bug fixed
A scoring system has been added, fiddling around with players and
bullets. We needed a system to add score to a player when a bullet kills
an enemy.

Also a bug was fixed. Enemies were able to shoot outside the boundary
before disappearing. By adding an external boundary this bug has been
fixed.